refactor: Product domain package move (tobe.domain -> domain)
