+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -1,16 +1,17 @@
-//package kitchenpos.products.domain;
-//
-//import java.util.List;
-//import java.util.Optional;
-//import java.util.UUID;
-//
-//public interface ProductRepository {
-//    Product save(Product product);
-//
-//    Optional<Product> findById(UUID id);
-//
-//    List<Product> findAll();
-//
-//    List<Product> findAllByIdIn(List<UUID> ids);
-//}
-//
+package kitchenpos.products.domain;
+
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface ProductRepository {
+
+    Product save(Product product);
+
+    Optional<Product> findById(UUID id);
+
+    List<Product> findAll();
+
+    List<Product> findAllByIdIn(List<UUID> ids);
+}
