+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,8 +1,8 @@
 package kitchenpos.products.application;
 
 
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
